feat: add modules data.js and util.js
